Make it so ALWAYS_VISIBLE applies to FoW but not to shroud
This Kind Of was originally only used for MultiplayerBeacon, but since the shroud is not used in multiplayer, I consider this a safe change. The problem with it making objects visible in the shroud is that this tends to look very awkward: the shroud should (and nearly always does) obscure absolutely everything.